requirement.txt
pytest~=7.1.2
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
pytest-base-url
allure-pytest
PyYAML~=6.0
requests
jsonpath
paramiko
redis
pymysql
mysqlclient